Refactor member area pages with modular components and async data fetching
