--- TODO_MEMBER_PAGES.md
@@ -1,46 +1,0 @@
-# Todo List: Member Pages Implementation
-
-## 1. Setup Member Area Structure
-- [x] Create layout.tsx for member-area with proper sidebar and authentication
-- [x] Create subdirectories for each page: consumi, benefici, documenti
-- [x] Extract sidebar navigation to a shared component (member-sidebar)
-
-## 2. "I miei Consumi" (My Consumption) Page
-- [x] Create app/member-area/consumi/page.tsx
-- [x] Create components/member-consumi.tsx component
-- [x] Include consumption charts (weekly, monthly, yearly views)
-- [x] Show energy consumption vs shared energy comparison
-- [x] Add export functionality for consumption data
-
-## 3. "Benefici Economici" (Economic Benefits) Page
-- [x] Create app/member-area/benefici/page.tsx
-- [x] Create components/member-benefici.tsx component
-- [x] Display savings overview cards
-- [x] Show incentives and savings charts
-- [x] Include historical benefits table
-- [x] Add projected savings calculator
-
-## 4. "Documenti" (Documents) Page
-- [x] Create app/member-area/documenti/page.tsx
-- [x] Create components/member-documenti.tsx component
-- [x] Implement document list with filters (type, date)
-- [x] Add download functionality
-- [x] Include document preview (mocked)
-- [x] Add search functionality
-
-## 5. Update Navigation
-- [x] Update links in member sidebar to point to new routes
-- [x] Ensure active state highlighting works correctly
-- [ ] Test navigation flow between pages
-
-## 6. Ensure Consistency
-- [x] Use the same styling patterns as existing pages
-- [x] Implement responsive design using existing components
-- [x] Follow the same authentication pattern
-- [x] Use mocked data following existing patterns
-
-## 7. Testing
-- [ ] Test all pages with member role
-- [ ] Verify authentication redirects work
-- [ ] Check responsive behavior on mobile/tablet
-- [ ] Ensure all interactive elements work (buttons, downloads, etc.)
