+++ app/member-area/benefici/loading.tsx
@@ -1,0 +1,5 @@
+import { BeneficiSkeleton } from "@/components/benefici-skeleton"
+
+export default function Loading() {
+  return <BeneficiSkeleton />
+}
