+++ app/member-area/consumi/loading.tsx
@@ -1,0 +1,5 @@
+import { ConsumiSkeleton } from "@/components/consumi-skeleton"
+
+export default function Loading() {
+  return <ConsumiSkeleton />
+}
