+++ app/member-area/documenti/loading.tsx
@@ -1,0 +1,5 @@
+import { DocumentiSkeleton } from "@/components/documenti-skeleton"
+
+export default function Loading() {
+  return <DocumentiSkeleton />
+}
