--- app/member-area/layout.tsx
@@ -1,129 +1,0 @@
-"use client"
-
-import { useEffect, useState } from "react"
-import { useRouter, usePathname } from "next/navigation"
-import { MemberSidebar } from "@/components/member-sidebar"
-import { Bell, Menu } from "lucide-react"
-import { Button } from "@/components/ui/button"
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuLabel,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu"
-import Image from "next/image"
-import { Sheet, SheetContent, SheetTrigger } from "@/components/ui/sheet"
-
-const getPageTitle = (pathname: string): string => {
-  const routes: Record<string, string> = {
-    "/member-area": "Dashboard",
-    "/member-area/consumi": "I Miei Consumi",
-    "/member-area/benefici": "Benefici Economici",
-    "/member-area/documenti": "Documenti",
-  }
-  return routes[pathname] || "Area Membri"
-}
-
-export default function MemberAreaLayout({
-  children,
-}: {
-  children: React.ReactNode
-}) {
-  const router = useRouter()
-  const pathname = usePathname()
-  const [user, setUser] = useState<any>(null)
-  const [isMobileMenuOpen, setIsMobileMenuOpen] = useState(false)
-
-  useEffect(() => {
-    // Check if user is authenticated and is a member
-    if (typeof window !== "undefined") {
-      const userData = localStorage.getItem("user")
-      if (userData) {
-        const parsedUser = JSON.parse(userData)
-        if (parsedUser.role !== "member") {
-          router.push("/login")
-          return
-        }
-        setUser(parsedUser)
-      } else {
-        router.push("/login")
-      }
-    }
-  }, [router])
-
-  const handleLogout = () => {
-    localStorage.removeItem("user")
-    router.push("/login")
-  }
-
-  if (!user) {
-    return <div className="flex h-screen items-center justify-center">Caricamento...</div>
-  }
-
-  return (
-    <div className="grid min-h-screen w-full md:grid-cols-[220px_1fr] lg:grid-cols-[280px_1fr]">
-      {/* Desktop Sidebar */}
-      <MemberSidebar />
-      
-      {/* Mobile Sidebar */}
-      <Sheet open={isMobileMenuOpen} onOpenChange={setIsMobileMenuOpen}>
-        <SheetContent side="left" className="p-0 w-[280px]">
-          <MemberSidebar />
-        </SheetContent>
-      </Sheet>
-
-      <div className="flex flex-col">
-        <header className="flex h-14 items-center gap-4 border-b bg-background px-4 lg:h-[60px] lg:px-6">
-          <Button
-            variant="outline"
-            size="icon"
-            className="md:hidden"
-            onClick={() => setIsMobileMenuOpen(true)}
-          >
-            <Menu className="h-5 w-5" />
-            <span className="sr-only">Toggle menu</span>
-          </Button>
-          
-          <div className="w-full flex-1">
-            <h1 className="text-lg font-semibold md:text-2xl">
-              {pathname === "/member-area" ? `Benvenuto, ${user?.name}` : getPageTitle(pathname)}
-            </h1>
-          </div>
-          
-          <Button variant="outline" size="icon" className="ml-auto h-8 w-8 bg-transparent">
-            <Bell className="h-4 w-4" />
-            <span className="sr-only">Toggle notifications</span>
-          </Button>
-          
-          <DropdownMenu>
-            <DropdownMenuTrigger asChild>
-              <Button variant="secondary" size="icon" className="rounded-full">
-                <Image
-                  src="/placeholder.svg?height=32&width=32"
-                  width={32}
-                  height={32}
-                  alt="Avatar"
-                  className="rounded-full"
-                />
-                <span className="sr-only">Toggle user menu</span>
-              </Button>
-            </DropdownMenuTrigger>
-            <DropdownMenuContent align="end">
-              <DropdownMenuLabel>Il Mio Account</DropdownMenuLabel>
-              <DropdownMenuSeparator />
-              <DropdownMenuItem>Impostazioni</DropdownMenuItem>
-              <DropdownMenuItem>Supporto</DropdownMenuItem>
-              <DropdownMenuSeparator />
-              <DropdownMenuItem onClick={handleLogout}>Logout</DropdownMenuItem>
-            </DropdownMenuContent>
-          </DropdownMenu>
-        </header>
-        <main className="flex flex-1 flex-col gap-4 p-4 md:gap-8 md:p-8">
-          {children}
-        </main>
-      </div>
-    </div>
-  )
-}
