+++ components/benefici/benefici-page-header.tsx
@@ -1,0 +1,7 @@
+export function BeneficiPageHeader() {
+  return (
+    <div>
+      <h2 className="text-3xl font-bold tracking-tight">Benefici Economici</h2>
+    </div>
+  )
+}
