+++ components/consumi/consumi-page-header.tsx
@@ -1,0 +1,7 @@
+export function ConsumiPageHeader() {
+  return (
+    <div>
+      <h2 className="text-3xl font-bold tracking-tight">I Miei Consumi</h2>
+    </div>
+  )
+}
