+++ components/documenti/documenti-page-header.tsx
@@ -1,0 +1,7 @@
+export function DocumentiPageHeader() {
+  return (
+    <div>
+      <h2 className="text-3xl font-bold tracking-tight">Documenti</h2>
+    </div>
+  )
+}
